detangle basic transforms
