+++ src/axis.js
@@ -1,7 +1,8 @@
 import {axisTop, axisBottom, axisRight, axisLeft, create, format, utcFormat} from "d3";
 import {formatIsoDate} from "./format.js";
-import {boolean, take, number, string, keyword, maybeKeyword, constant, isTemporal} from "./mark.js";
+import {boolean, take, number, string, keyword, maybeKeyword, isTemporal} from "./mark.js";
 import {radians} from "./math.js";
+import {constant} from "./options.js";
 import {impliedString} from "./style.js";
 
 export class AxisX {
