+++ src/index.js
@@ -1,5 +1,5 @@
 export {plot} from "./plot.js";
-export {Mark, marks, valueof} from "./mark.js";
+export {Mark, marks} from "./mark.js";
 export {Area, area, areaX, areaY} from "./marks/area.js";
 export {BarX, BarY, barX, barY} from "./marks/bar.js";
 export {Cell, cell, cellX, cellY} from "./marks/cell.js";
@@ -13,9 +13,8 @@ export {RuleX, RuleY, ruleX, ruleY} from "./marks/rule.js";
 export {Text, text, textX, textY} from "./marks/text.js";
 export {TickX, TickY, tickX, tickY} from "./marks/tick.js";
 export {Vector, vector} from "./marks/vector.js";
-export {filter} from "./transforms/filter.js";
-export {reverse} from "./transforms/reverse.js";
-export {sort, shuffle} from "./transforms/sort.js";
+export {valueof} from "./options.js";
+export {filter, reverse, sort, shuffle} from "./transforms/basic.js";
 export {bin, binX, binY} from "./transforms/bin.js";
 export {group, groupX, groupY, groupZ} from "./transforms/group.js";
 export {normalize, normalizeX, normalizeY} from "./transforms/normalize.js";
