+++ src/mark.js
@@ -1,13 +1,10 @@
 import {ascending, color, descending, rollup, sort} from "d3";
+import {arrayify, isOptions, valueof} from "./options.js";
 import {plot} from "./plot.js";
 import {registry} from "./scales/index.js";
 import {styles} from "./style.js";
 import {basic} from "./transforms/basic.js";
 import {maybeReduce} from "./transforms/group.js";
-
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/TypedArray
-const TypedArray = Object.getPrototypeOf(Uint8Array);
-const objectToString = Object.prototype.toString;
 
 export class Mark {
   constructor(data, channels = [], options = {}, defaults) {
@@ -105,17 +102,6 @@ function channelSort(channels, facetChannels, data, options) {
   }
 }
 
-// This allows transforms to behave equivalently to channels.
-export function valueof(data, value, type) {
-  const array = type === undefined ? Array : type;
-  return typeof value === "string" ? array.from(data, field(value))
-    : typeof value === "function" ? array.from(data, value)
-    : typeof value === "number" || value instanceof Date ? array.from(data, constant(value))
-    : value && typeof value.transform === "function" ? arrayify(value.transform(data), type)
-    : arrayify(value, type); // preserve undefined type
-}
-
-export const field = name => d => d[name];
 export const indexOf = (d, i) => i;
 export const identity = {transform: d => d};
 export const zero = () => 0;
@@ -124,7 +110,6 @@ export const number = x => x == null ? x : +x;
 export const boolean = x => x == null ? x : !!x;
 export const first = d => d[0];
 export const second = d => d[1];
-export const constant = x => () => x;
 
 // A few extra color keywords not known to d3-color.
 const colors = new Set(["currentColor", "none"]);
@@ -160,27 +145,6 @@ export function keyword(input, name, allowed) {
   const i = `${input}`.toLowerCase();
   if (!allowed.includes(i)) throw new Error(`invalid ${name}: ${input}`);
   return i;
-}
-
-// Promotes the specified data to an array or typed array as needed. If an array
-// type is provided (e.g., Array), then the returned array will strictly be of
-// the specified type; otherwise, any array or typed array may be returned. If
-// the specified data is null or undefined, returns the value as-is.
-export function arrayify(data, type) {
-  return data == null ? data : (type === undefined
-    ? (data instanceof Array || data instanceof TypedArray) ? data : Array.from(data)
-    : (data instanceof type ? data : type.from(data)));
-}
-
-// Disambiguates an options object (e.g., {y: "x2"}) from a primitive value.
-export function isObject(option) {
-  return option && option.toString === objectToString;
-}
-
-// Disambiguates an options object (e.g., {y: "x2"}) from a channel value
-// definition expressed as a channel transform (e.g., {transform: …}).
-export function isOptions(option) {
-  return isObject(option) && typeof option.transform !== "function";
 }
 
 // For marks specified either as [0, x] or [x1, x2], such as areas and bars.
