--- src/transforms/reverse.js
@@ -1,9 +1,0 @@
-import {basic} from "./basic.js";
-
-export function reverse(options) {
-  return basic(options, reverseTransform);
-}
-
-export function reverseTransform(data, facets) {
-  return {data, facets: facets.map(I => I.slice().reverse())};
-}
